03_Optimization/requirements.txt
dspy-ai
litellm
pydantic
pytest
